backend/minimal_requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
aiofiles
numpy
